+++ packages/world-vercel/src/queue.ts
@@ -6,6 +6,7 @@ import {
   ValidQueueName,
 } from '@workflow/world';
 import * as z from 'zod';
+import { type APIConfig, getHeaders, getHttpUrl } from './utils.js';
 
 const MessageWrapper = z.object({
   payload: QueuePayloadSchema,
@@ -14,7 +15,25 @@ const MessageWrapper = z.object({
 
 const VERCEL_QUEUE_MAX_VISIBILITY = 82800; // 23 hours in seconds
 
-export function createQueue(): Queue {
+export function createQueue(config?: APIConfig): Queue {
+  const { baseUrl, usingProxy } = getHttpUrl(config);
+  const headers = getHeaders(config);
+  if (usingProxy) {
+    // If we're using a proxy for the Workflow API, we should also go
+    // through the proxy for the queues API.
+    process.env.VERCEL_QUEUE_BASE_URL = `${baseUrl}`;
+    process.env.VERCEL_QUEUE_BASE_PATH = '/queues/v2/messages';
+    if (config?.token) {
+      process.env.VERCEL_QUEUE_TOKEN = config.token;
+    }
+    if (headers) {
+      headers.forEach((value, key) => {
+        const sanitizedKey = key.replaceAll('-', '__');
+        process.env[`VERCEL_QUEUE_HEADER_${sanitizedKey}`] = value;
+      });
+    }
+  }
+
   const queue: Queue['queue'] = async (queueName, x, opts) => {
     const encoded = MessageWrapper.encode({
       payload: x,