+++ packages/web/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @workflow/web
 
+## 4.0.1-beta.5
+
+### Patch Changes
+
+- 0f845af: Alias workflow web to workflow inspect runs --web, hide trace viewer search for small runs
+- ffb7af3: Web: make error handling local/inline to where it's used, unify API error responses
+
 ## 4.0.1-beta.4
 
 ### Patch Changes